Move idle activity state into core
IdleState is the shared runtime activity signal for MCP, CDP, watchdogs, and persistence, so give it a core activity owner instead of leaving it in generic utilities.

Constraint: Preserve the existing idle-adaptive monitor and timeout behavior.

Rejected: Keep idle-state in src/utils | It is a cross-cutting runtime state primitive with a clear architecture role.

Confidence: high

Scope-risk: moderate

Directive: Shared runtime activity signals should live under src/core/activity.

Tested: npm test -- --runTestsByPath tests/core/activity/idle-state.test.ts tests/core/process/idle-timeout.test.ts tests/watchdog/idle-adaptive-monitors.test.ts tests/integration/idle-adaptive-monitoring.test.ts tests/integration/embedded-server.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check

Not-tested: Full manual workflow_dispatch test matrix.
