+++ src/core/activity/idle-state.ts
