+++ src/core/server.ts
@@ -42,7 +42,7 @@ import { getLifecycleBus, type LifecycleEventBus } from './lifecycle';
 import { writePidFile, cleanOrphanedChromeProcesses } from '../chrome/pid-manager';
 import { installParentWatcher } from './process/parent-watcher';
 import { installIdleTimeout } from './process/idle-timeout';
-import { getIdleState } from '../utils/idle-state';
+import { getIdleState } from './activity/idle-state';
 import { getListenerErrorStats } from '../utils/safe-listener';
 import type { MCPServer } from '../mcp-server';
 import {
