+++ src/watchdog/disk-monitor.ts
@@ -10,7 +10,7 @@
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import * as os from 'os';
-import { getIdleState, IDLE_WINDOW_MS, IdleState } from '../utils/idle-state';
+import { getIdleState, IDLE_WINDOW_MS, IdleState } from '../core/activity/idle-state';
 
 /** Idle cadence. 30 min is 6× slower than the default 5 min active rate. */
 const IDLE_INTERVAL_MS = 30 * 60_000;
