+++ tests/integration/idle-adaptive-monitoring.test.ts
@@ -25,7 +25,7 @@ import { ChromeProcessMonitor } from '../../src/watchdog/chrome-monitor';
 import { DiskMonitor } from '../../src/watchdog/disk-monitor';
 import { BrowserStateManager } from '../../src/browser-state/snapshot';
 import { SessionStatePersistence } from '../../src/session/state-persistence';
-import { createIdleState, IDLE_WINDOW_MS } from '../../src/utils/idle-state';
+import { createIdleState, IDLE_WINDOW_MS } from '../../src/core/activity/idle-state';
 
 describe('idle-adaptive monitoring integration (issue #649 §5.5)', () => {
   const fs = require('fs');
